refactor: adjusting api structure for better performance
