+++ app/api/products/[id]/route.ts
@@ -1,4 +1,5 @@
-import { makeFindProductByIdService } from '@/src/services/factories';
+
+import { makeFindProductByIdService } from '@/src/presentation/factory/services/products/find-product-by-id-factory';
 import { NextResponse, NextRequest } from 'next/server';
 
 export async function GET(request: NextRequest, { params }: { params: { id: string } }) {
