--- src/services/factories/index.ts
@@ -1,1 +1,0 @@
-export * from './find-product-by-id-factory';
